Migrate pipeline transforms to compute module
Move core logic from pipeline/ to compute/. Update unit/integration
tests imports.
