+++ platform/vision/src/galadril_vision/compute/BUILD.bazel
@@ -1,0 +1,25 @@
+load("@rules_python//python:defs.bzl", "py_library")
+
+py_library(
+    name = "compute",
+    srcs = [
+        "__init__.py",
+        "helpers.py",
+        "loader.py",
+        "runtime.py",
+        "tasks.py",
+    ],
+    imports = ["../.."],
+    visibility = [
+        "//platform/vision/src:__subpackages__",
+        "//platform/vision/tests:__subpackages__",
+    ],
+    deps = [
+        "//platform/vision/src/galadril_vision/common",
+        "//platform/vision/src/galadril_vision/connectors/postgres",
+        "@pip//numpy",
+        "@pip//opencv_python_headless",
+        "@pip//orjson",
+        "@pip//structlog",
+    ],
+)
