+++ platform/vision/src/galadril_vision/compute/helpers.py
