+++ platform/vision/src/galadril_vision/compute/loader.py
