+++ platform/vision/src/galadril_vision/pipeline/BUILD.bazel
@@ -4,16 +4,11 @@ py_library(
     name = "pipeline",
     srcs = [
         "__init__.py",
-        "dagster_client.py",
-        "dagster_defs.py",
+        "client.py",
+        "defs.py",
         "executor.py",
-        "model_loader.py",
-        "postgres_tasks.py",
         "router.py",
         "runner.py",
-        "transform_helpers.py",
-        "transforms.py",
-        "worker_runtime.py",
     ],
     imports = ["../.."],
     visibility = [
@@ -21,12 +16,10 @@ py_library(
         "//platform/vision/tests:__subpackages__",
     ],
     deps = [
-        "//machine-learning/galadril-inference/src/galadril_inference/common",
-        "//machine-learning/galadril-inference/src/galadril_inference/core",
-        "//machine-learning/galadril-inference/src/galadril_inference/storage",
         "//platform/pipeline/src/galadril_pipeline",
         "//platform/vision/src/galadril_vision/causal",
         "//platform/vision/src/galadril_vision/common",
+        "//platform/vision/src/galadril_vision/compute/udfs",
         "//platform/vision/src/galadril_vision/connectors/kafka",
         "//platform/vision/src/galadril_vision/connectors/postgres",
         "//platform/vision/src/galadril_vision/connectors/s3",
@@ -35,9 +28,6 @@ py_library(
         "@pip//boto3",
         "@pip//daft",
         "@pip//dagster",
-        "@pip//numpy",
-        "@pip//opencv_python_headless",
-        "@pip//orjson",
         "@pip//pydantic",
         "@pip//pyyaml",
         "@pip//structlog",
