+++ platform/vision/src/galadril_vision/pipeline/__init__.py
@@ -3,19 +3,9 @@
 from galadril_vision.pipeline.executor import ESKGPipelineExecutor
 from galadril_vision.pipeline.router import MultiTenantPipelineRouter
 from galadril_vision.pipeline.runner import VisionPipeline
-from galadril_vision.pipeline.transforms import (
-    DownloadDataWorker,
-    resolve_entities_udf,
-    run_inference_udf,
-    sink_to_db_udf,
-)
 
 __all__ = [
     "ESKGPipelineExecutor",
     "MultiTenantPipelineRouter",
     "VisionPipeline",
-    "DownloadDataWorker",
-    "run_inference_udf",
-    "resolve_entities_udf",
-    "sink_to_db_udf",
 ]
