+++ platform/vision/src/galadril_vision/pipeline/client.py
