+++ platform/vision/src/galadril_vision/pipeline/defs.py
