+++ platform/vision/tests/BUILD.bazel
@@ -6,6 +6,7 @@ py_test(
     deps = [
         "//platform/pipeline/src/galadril_pipeline",
         "//platform/vision/src/galadril_vision/common",
+        "//platform/vision/src/galadril_vision/compute",
         "//platform/vision/src/galadril_vision/connectors/kafka",
         "//platform/vision/src/galadril_vision/pipeline",
         "@pip//aioboto3",
@@ -33,8 +34,9 @@ py_test(
     name = "test_pipeline_transforms",
     srcs = ["test_pipeline_transforms.py"],
     deps = [
+        "//platform/vision/src/galadril_vision/compute",
+        "//platform/vision/src/galadril_vision/compute/udfs",
         "//platform/vision/src/galadril_vision/connectors/postgres",
-        "//platform/vision/src/galadril_vision/pipeline",
         "@pip//boto3",  # keep
     ],
 )
@@ -42,7 +44,7 @@ py_test(
 py_test(
     name = "test_transforms",
     srcs = ["test_transforms.py"],
-    deps = ["//platform/vision/src/galadril_vision/pipeline"],
+    deps = ["//platform/vision/src/galadril_vision/compute"],
 )
 
 py_test(
@@ -64,7 +66,7 @@ py_test(
 py_test(
     name = "test_postgres_tasks",
     srcs = ["test_postgres_tasks.py"],
-    deps = ["//platform/vision/src/galadril_vision/pipeline"],
+    deps = ["//platform/vision/src/galadril_vision/compute"],
 )
 
 py_test(
